# docs/swagger.yaml
basePath: /
definitions:
...
  handler.entryForm:
    properties:
      confirm_password:
        type: string
      email:
        type: string
      password:
        type: string
    type: object
...
  models.AgeCategory:
    properties:
      id:
        type: integer
      max_age:
        type: integer
      min_age:
        type: integer
    type: object
  models.Cover:
    properties:
      filename:
        type: string
      id:
        type: integer
      project_id:
        type: integer
    type: object
...
  models.Genre:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.Keyword:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.Movie:
    properties:
      age_categories:
        items:
          $ref: '#/definitions/models.AgeCategory'
        type: array
      cover:
        $ref: '#/definitions/models.Cover'
      description:
        type: string
      director:
        type: string
      duration:
        type: integer
      genres:
        items:
          $ref: '#/definitions/models.Genre'
        type: array
      id:
        type: integer
      keywords:
        items:
          $ref: '#/definitions/models.Keyword'
        type: array
      popularity:
        type: integer
      producer:
        type: string
      release_year:
        type: integer
      screenshots:
        items:
          $ref: '#/definitions/models.Screenshot'
        type: array
      title:
        type: string
      youtube_id:
        type: string
    type: object
...
  models.Screenshot:
    properties:
      filename:
        type: string
      id:
        type: integer
      project_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Type "JWT" token received from the server
  title: Ozinshe API
  version: "1.0"
paths:
  /ages:
    get:
      description: Retrieves a list of all age categories in the system.
      produces:
      - application/json
      responses:
        "200":
          description: List of age categories
          schema:
            items:
              $ref: '#/definitions/models.AgeCategory'
            type: array
        "400":
          description: Error retrieving age categories
          schema: {}
      summary: Get a list of all age categories
      tags:
      - age categories
    post:
      consumes:
      - multipart/form-data
      description: Creates a new age category based on the provided age range. Requires
        admin authorization.
      parameters:
      - description: Age range (e.g., '18-30')
        in: formData
        name: age
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Age category created
        "400":
          description: Error creating age category
          schema: {}
      security:
      - CookieAuth: []
      summary: Create a new age category
      tags:
      - age categories
  /ages/{id}:
    delete:
      description: Deletes an existing age category based on the provided ID.Requires
        admin authorization.
      parameters:
      - description: Age category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Age category deleted successfully
        "400":
          description: Error deleting age category
          schema: {}
      security:
      - CookieAuth: []
      summary: Delete an existing age category
      tags:
      - age categories
    get:
      description: Retrieves details of an age category based on the provided ID.
      parameters:
      - description: Age category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Age category retrieved successfully
          schema:
            $ref: '#/definitions/models.AgeCategory'
        "400":
          description: Error retrieving age category
          schema: {}
      summary: Get details of a specific age category
      tags:
      - age categories
...
  /genres:
    get:
      description: Retrieves a list of all genres.
      produces:
      - application/json
      responses:
        "200":
          description: List of genres
          schema:
            items:
              $ref: '#/definitions/models.Genre'
            type: array
        "400":
          description: Error getting genres
          schema: {}
      summary: Get a list of all genres
      tags:
      - genres
    post:
      consumes:
      - multipart/form-data
      description: Creates a new genre with the provided name. Requires admin authorization.
      parameters:
      - description: Genre name
        in: formData
        name: genre
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Genre created
        "400":
          description: Error creating genre
          schema: {}
      security:
      - CookieAuth: []
      summary: Create a new genre
      tags:
      - genres
  /genres/{id}:
    delete:
      description: Deletes an existing genre based on the provided ID. Requires admin
        authorization.
      parameters:
      - description: Genre ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Genre deleted successfully
        "400":
          description: Error deleting genre
          schema: {}
      security:
      - CookieAuth: []
      summary: Delete an existing genre
      tags:
      - genres
    get:
      description: Retrieves details of a genre based on the provided ID.
      parameters:
      - description: Genre ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Genre details
          schema:
            $ref: '#/definitions/models.Genre'
        "400":
          description: Error getting genre
          schema: {}
      summary: Get details of a specific genre
      tags:
      - genres
  /movies:
    get:
      description: Retrieves a list of all movies.
      produces:
      - application/json
      responses:
        "200":
          description: List of movies
          schema:
            items:
              $ref: '#/definitions/models.Movie'
            type: array
        "400":
          description: Error getting movies
          schema: {}
      summary: Get a list of all movies
      tags:
      - movies
  /movies/{id}:
    delete:
      description: Deletes an existing movie based on the provided ID. Requires admin
        authorization.
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Movie deleted successfully
        "500":
          description: Error deleting movie
          schema: {}
      security:
      - CookieAuth: []
      summary: Delete an existing movie
      tags:
      - movies
    get:
      description: Retrieves details of a movie based on the provided ID.
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Movie details
          schema:
            $ref: '#/definitions/models.Movie'
        "400":
          description: Error getting movie
          schema: {}
      summary: Get details of a specific movie
      tags:
      - movies
...
  /signin:
    post:
      consumes:
      - application/json
      description: Logs in an existing user with the provided email and password.
      parameters:
      - description: User email and password
        in: body
        name: entry
        required: true
        schema:
          $ref: '#/definitions/handler.entryForm'
      produces:
      - application/json
      responses:
        "200":
          description: User logged in successfully
          schema:
            items:
              type: string
            type: array
        "400":
          description: 'Error signing in: user loggin in failed'
          schema:
            type: string
      summary: User sign in
      tags:
      - authentication
  /signout:
    get:
      description: Logs out the currently authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
        "400":
          description: Error logging out user
          schema: {}
      security:
      - CookieAuth: []
      summary: User sign out
      tags:
      - authentication
  /signup:
    post:
      consumes:
      - application/json
      description: Registers a new user with the provided email and password.
      parameters:
      - description: User email and password
        in: body
        name: entry
        required: true
        schema:
          $ref: '#/definitions/handler.entryForm'
      produces:
      - application/json
      responses:
        "200":
          description: User registered successfully
        "400":
          description: 'Error signing up: user registration failed'
          schema:
            type: string
      summary: User sign up
      tags:
      - authentication
...
swagger: "2.0"